requirements.txt
torch>=2.1
transformers>=4.40
datasets>=2.18
accelerate>=0.28
peft>=0.10
trl>=0.8
bitsandbytes>=0.43
wandb
matplotlib
pandas
seaborn
scikit-learn
sentencepiece
spacy
jupyter
ipykernel
huggingface_hub
python-dotenv